requirements.txt
Flask==2.2.2
Flask-Mail
Flask-WTF
psycopg2-binary
supabase
python-dotenv
gunicorn